% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwqs_stan_models.R
\docType{data}
\name{bwqs_covmod_binY}
\alias{bwqs_covmod_binY}
\title{STAN definitions of BWQS models for binary outcome Y.}
\format{
An object of class \code{character} of length 1.
}
\usage{
bwqs_covmod_binY
}
\description{
STAN definitions of BWQS models for binary outcome Y.
}
\keyword{datasets}
