% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dbwqs.R
\name{ccomp_plot}
\alias{ccomp_plot}
\title{DBWQS change plot.
Composition proportions distribution plot.}
\usage{
ccomp_plot(ccomp, sort_by = NULL, decreasing = TRUE)
}
\arguments{
\item{ccomp}{data frame containing the compositional variables.}

\item{sort_by}{Character variable representing the name of the variable to sort by in \code{ccomp}. Default value is \code{NULL} (no sorting).}

\item{decreasing}{\code{TRUE/FALSE}: should the sorting be decreasing or increasing? Default is \code{TRUE}. Only considered when \code{sort_by} is not \code{NULL}.}
}
\value{
ggplot object containing the composition proportions distribution plot.
}
\description{
Construct stacked bar plot representing the distribution of a given compositional variable.
}
