% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dbwqs.R
\name{dbwqs}
\alias{dbwqs}
\title{Dirichlet-Bayesian Weighted Quantile Sum regression models.}
\usage{
dbwqs(
  data,
  out,
  exp,
  cov = NULL,
  ref = NULL,
  q = 4,
  chains = 1,
  impute_zeros = F,
  eps = 1e-04,
  seed = 99221,
  warmup = 1000,
  iter = 2000,
  thin = 1,
  verbose = F,
  max_treedepth = 20,
  adapt_delta = 0.999999
)
}
\arguments{
\item{data}{Data frame containing outcomes, exposures and covariates.}

\item{out}{Character vector of compositional outcome names corresponding to those in \code{data}.}

\item{exp}{Character vector of exposure names included in the mixture, corresponding to those in \code{data}.}

\item{cov}{Character vector covariate names, corresponding to those in \code{data}. Note that the categorical covariates need to be included as factors in \code{data}; otherwise, they will be treated as numeric in the model.}

\item{ref}{Character specifying the name of the outcome variable to be used as reference for the multinomial logit link. By default (\code{ref=NULL}), the first outcome variable in \code{out} is taken to be the reference.}

\item{q}{Integer specifying how exposure variables will be ranked, e.g. in quartiles
(\code{q = 4}), deciles (\code{q = 10}), or percentiles (\code{q = 100}). Default is \code{q=4}.}

\item{chains}{Integer specifying the number of chains in Hamiltonian Monte Carlo algorithm.
Default value \code{chains = 1}.}

\item{impute_zeros}{Boolean value indicating if zero values in the compositional outcomes should be imputed. Default is \code{impute_zeros=FALSE}.}

\item{eps}{Numeric value defining a small threshold below which the outcome proportion will be rounded to zero. Default value is \code{eps=0.0001}. This parameter is only used when \code{impute_zeros=TRUE}.}

\item{seed}{Integer value to fix the seed for reproducibility of results.}

\item{warmup}{Integer spcifying the number of warmup iterations in HMC algorithm. Default value is \code{warmup = 2000}.}

\item{iter}{Integer specifying the number of iterations in HMC algorithm. Default value is \code{iter = 2000}. This value is chosen to make testing faster; it is recommended to use 10000 iterations in a full analysis.}

\item{thin}{Integer specifying the thinning parameter in Hamiltonian Monte Carlo algorithm. Default value is \code{thin=1}.}

\item{verbose}{Boolean value indicating if user wants to display additional information during fitting. Default value is \code{verbose=FALSE}.}

\item{max_treedepth}{Integer specifying maximum tree depth for HMC algorithm. Default value is \code{max_treedepth=20}.}

\item{adapt_delta}{Numeric value specifying the target average acceptance probability for HMC algorithm. Default value is \code{adapt_delta = 0.999999}.}
}
\value{
\code{dbwqs} returns a list containing the following elements:
\item{dbwqs_fit}{An \code{S4} object with all details of the Hamiltonian Monte Carlo, all the extractions
from the posterior distribution and all values of the parameters}
\item{absprop}{Table with the statistics of the mixture index coefficients in absolute proportion scale: mean, standard error of the mean,
standard deviation, lower and upper values for the 80\% and 95\% credible interval, n_eff and Rhat.}
\item{relprop}{Table with the statistics of the mixture index coefficients in relative proportion scale: mean, standard error of the mean,
standard deviation, lower and upper values for the 80\% and 95\% credible interval, n_eff and Rhat.}
\item{mixprop}{Table with the statistics of the mixture weights: mean, standard error of the mean,
standard deviation, lower and upper values for the 80\% and 95\% credible interval, n_eff and Rhat.}
\item{phi}{Table with the statistics of the precision parameter: mean, standard error of the mean,
standard deviation, lower and upper values for the 80\% and 95\% credible interval, n_eff and Rhat.}
\item{error_flag}{Flag indicating whether the fitting process was successful (0:success / -1: failed).}
}
\description{
Fits the Dirichlet Bayesian Weighted Quantile Sum (DBWQS) regression for compositional outcomes. The function \code{dbwqs} uses the package \code{rstan} which allows the connection with the STAN software written in C++ for bayesian inference; for further information, please consult https://mc-stan.org/.
}
\author{
Hachem Saddiki
}
