% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dbwqs.R
\name{dbwqs_acplot}
\alias{dbwqs_acplot}
\title{Auto-correlation plots from fitted DBWQS model.}
\usage{
dbwqs_acplot(dbwqs_obj)
}
\arguments{
\item{dbwqs_obj}{Object containing the fitted DBWQS model (from \code{dbwqs} output).}
}
\value{
Returns a list containins 4 plot objects:
\itemize{
\item \code{relcoef_acplot}: Auto-correlation plots for coefficients associated with the exposure mixture index in relative proportion scale.
\item \code{abscoef_acplot}: Auto-correlation plots for coefficients associated with the exposure mixture index in absolute proportion scale.
\item \code{mixweights_acplot}: Auto-correlation plots for the estimated exposure mixture weights.
\item \code{phi_acplot}: Auto-correlation plot for the precision parameter \code{phi}.
}
}
\description{
Construct auto-correlation plots using R-STAN plotting functions
}
