% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbwqs_stan_models.R
\docType{data}
\name{dbwqs_covmod}
\alias{dbwqs_covmod}
\title{STAN definitions for Dirichlet-Bayesian Weighted Quantile Sum regression models.}
\format{
An object of class \code{character} of length 1.
}
\usage{
dbwqs_covmod
}
\description{
STAN definitions for Dirichlet-Bayesian Weighted Quantile Sum regression models.
}
\keyword{datasets}
