% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dbwqs.R
\name{dbwqs_traceplot}
\alias{dbwqs_traceplot}
\title{Trace plots from fitted DBWQS model.}
\usage{
dbwqs_traceplot(dbwqs_obj)
}
\arguments{
\item{dbwqs_obj}{Object containing the fitted DBWQS model (from \code{dbwqs} output).}
}
\value{
Returns a list containins 4 plot objects:
\itemize{
\item \code{relcoef_traceplot}: Trace plots for coefficients associated with the exposure mixture index in relative proportion scale.
\item \code{abscoef_traceplot}: Trace plots for coefficients associated with the exposure mixture index in absolute proportion scale.
\item \code{mixweights_traceplot}: Trace plots for the estimated exposure mixture weights.
\item \code{phi_traceplot}: Trace plot for the precision parameter \code{phi}.
}
}
\description{
Construct trace plots using R-STAN plotting functions.
}
