% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbwqs_simdata.R
\name{gen_dbwqs_simdata}
\alias{gen_dbwqs_simdata}
\title{Generate simulated compositional regression data set with exposure mixtures, covariates, and compositional outcomes.}
\usage{
gen_dbwqs_simdata(
  thetas = NULL,
  betas = NULL,
  wt = NULL,
  N = 200,
  C = 3,
  K = 4,
  J = 2,
  rho = 0.5,
  phi = 5,
  q = 4,
  seed = 99221
)
}
\arguments{
\item{thetas}{Numeric vector of size \code{C} representing the true regression coefficients of the exposure mixture index. The first element of \code{thetas} is equal to 0 by definition. If \code{NULL} (default), coefficients are randomly drawn from a uniform distribution.}

\item{betas}{Numeric matrix of size \code{J} by \code{C} representing the true regression coefficients of the simulated covariates. The first column of \code{betas} is equal to 0 by definition. If \code{NULL} (default), coefficients are randomly drawn from a uniform distribution.}

\item{wt}{Numeric vector of size \code{K} representing the exposure mixture weights. These weights should all be positive and sum to 1. If \code{NULL} (default), weights are randomly drawn from a Dirichlet distribution.}

\item{N}{Integer specifying the number of simulated observations (default=200).}

\item{C}{Integer specifying the number of components in the simulated compositional outcomes (default=3).}

\item{K}{Integer specifying the number of exposures included in the simulated exposure mixture (default=4).}

\item{J}{Integer specifying the number of simulated covariates (default = 2).}

\item{rho}{Numeric value in (0,1) specifying the correlation between simulated exposures (default=0.5).}

\item{phi}{Numeric value greater than 0 specifying the true precision parameter for the DBWQS model (default=5).}

\item{q}{Integer defining the exposure quantile split (default = 4).}

\item{seed}{Integer specifying seed value for reproducibility.}
}
\value{
This function returns a list containing the following:
\item{obsdata}{Data frame of simulated observations of exposures, covariates, and compositional outcomes.}
\item{thetas}{Vector of true effect estimates associated with the exposure mixture index in relative proportion scale.}
\item{absprops_per}{Vector of true effect estimates associated with exposure mixture index in absolute proportion percentage scale}
\item{wt}{Vector of true exposure mixture weights used to simulate the exposue mixture index.}
\item{betas}{Matrix of true regression coefficients associated with the covariates.}
}
\description{
Generate simulated compositional regression data set with exposure mixtures, covariates, and compositional outcomes.
}
\author{
Hachem Saddiki
}
