% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE75248_wCellComp}
\alias{GSE75248_wCellComp}
\title{A subset of data from the Rhode Island Child Health Study (RICHS)}
\format{
\subsection{GSE75248_wCellComp}{

Six estimated placenta cell type proportions (Syncytiotrophoblast,
trophoblast, stromal, nucleated red blood cells (nRBC), Hofbauer, and
endothelial cells). The exposure mixture is placenta metals labelled as
“placenta_as”, “placenta_pb”, etc. The rest are covariates defined
in the directory where the data was retrieved (see source).
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75248}
}
\usage{
GSE75248_wCellComp
}
\description{
A subset of data from the Rhode Island Child Health Study (RICHS)
}
\keyword{datasets}
