% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbwqs_stan_models.R
\docType{data}
\name{hbwqs_covmod_contY}
\alias{hbwqs_covmod_contY}
\title{STAN definitions for Hierarchical BWQS (HBWQS) with continuous outcome Y.}
\format{
An object of class \code{character} of length 1.
}
\usage{
hbwqs_covmod_contY
}
\description{
STAN definitions for Hierarchical BWQS (HBWQS) with continuous outcome Y.
}
\keyword{datasets}
