% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbwqs_stan_models.R
\docType{data}
\name{hbwqs_covmod_countY}
\alias{hbwqs_covmod_countY}
\title{STAN definitions of Hierarchical BWQS (HBWQS) models for count outcome Y.}
\format{
An object of class \code{character} of length 1.
}
\usage{
hbwqs_covmod_countY
}
\description{
STAN definitions of Hierarchical BWQS (HBWQS) models for count outcome Y.
}
\keyword{datasets}
