% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{impute_zeros}
\alias{impute_zeros}
\title{Impute zero proportions in compositional outcomes with an estimated lower limit of detection}
\usage{
impute_zeros(df, eps = 1e-04)
}
\arguments{
\item{df}{Data frame containing compositional outcomes, one for each column.}

\item{eps}{Numeric value defining a small threshold below which the outcome proportion will be rounded to zero (\code{default=1e-4}).}
}
\value{
Data frame containing compositional outcomes where all zeros are imputed with an estimated lower limit of detection.
}
\description{
Impute zero proportions in compositional outcomes with an estimated lower limit of detection
}
