% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{quantile_split}
\alias{quantile_split}
\title{Quantize one or multiple continuous variables using ecdf.}
\usage{
quantile_split(data, mix_name = NULL, q = 4)
}
\arguments{
\item{data}{Data frame with at least one continuous variable.}

\item{mix_name}{Character vector containing names of variable to quantize. If null, all variables in \code{data} are quantized.}

\item{q}{Integer defining the quantile split (default = 4).}
}
\value{
Data frame with original variables replaced by their quantized version.
}
\description{
Quantize one or multiple continuous variables using ecdf.
}
